suppress MSVC warning

+++ include/cinolib/dual_mesh.cpp
@@ -97,7 +97,7 @@ void dual_mesh(const AbstractPolyhedralMesh<M,V,E,F,P> & primal,
         if(n_creases> 2)
         {
             crease_corner.at(vid) = true;
-            pv2dv[vid] = dual_verts.size();
+            pv2dv[vid] = uint(dual_verts.size());
             dual_verts.push_back(primal.vert(vid));
         }
     }
@@ -106,7 +106,7 @@ void dual_mesh(const AbstractPolyhedralMesh<M,V,E,F,P> & primal,
     {
         if(primal.edge_is_on_srf(eid) && primal.edge_data(eid).flags[CREASE])
         {
-            pe2dv[eid] = dual_verts.size();
+            pe2dv[eid] = uint(dual_verts.size());
             dual_verts.push_back(primal.edge_sample_at(eid, 0.5));
         }
     }
@@ -115,7 +115,7 @@ void dual_mesh(const AbstractPolyhedralMesh<M,V,E,F,P> & primal,
     {
         if(primal.face_is_on_srf(fid))
         {
-            pf2dv[fid] = dual_verts.size();
+            pf2dv[fid] = uint(dual_verts.size());
             dual_verts.push_back(primal.face_centroid(fid));
         }
     }
@@ -223,7 +223,7 @@ void dual_mesh(const AbstractPolyhedralMesh<M,V,E,F,P> & primal,
             auto query = f_map.find(key);
             if(query == f_map.end())
             {
-                uint fresh_id = dual_faces.size();
+                uint fresh_id = uint(dual_faces.size());
                 f_map[key] = fresh_id;
                 dual_faces.push_back(face);
                 poly.push_back(fresh_id);
@@ -286,7 +286,7 @@ void dual_mesh(const AbstractPolygonMesh<M,V,E,P>   & primal,
     {
         if(primal.vert_is_boundary(vid))
         {
-            v2verts[vid] = dual_verts.size();
+            v2verts[vid] = uint(dual_verts.size());
             dual_verts.push_back(primal.vert(vid));
         }
     }
@@ -294,7 +294,7 @@ void dual_mesh(const AbstractPolygonMesh<M,V,E,P>   & primal,
     {
         if(primal.edge_is_boundary(eid))
         {
-            e2verts[eid] = dual_verts.size();
+            e2verts[eid] = uint(dual_verts.size());
             dual_verts.push_back(primal.edge_sample_at(eid, 0.5));
         }
     }
